+++ src/components/Intro/IntroHeader/style.ts
