+++ src/components/Intro/LocationCard/index.tsx
