+++ src/components/Intro/LocationCard/style.ts
