+++ src/components/Intro/LogoDescription/index.tsx
