+++ src/components/Intro/LogoDescription/style.ts
