+++ src/components/Intro/VeganDescription/index.tsx
