+++ src/components/Intro/VeganDescription/style.ts
