+++ src/components/Intro/WayToCome/index.tsx
