+++ src/components/Intro/WayToCome/style.ts
